Move new helper functions to new file pod.go
